Fix regression of .OP without arguments (to be pushed upstream).
When the first argument is an empty string, groff-1.23.0 wrongly sets
the closing bracket in bold font.

This got recklessly broken in this commit:

groff commit c0cfb88b gbranden@ Nov 9, 2021
tmac/an.tmac: Fix code style nits.

It is hard to understand how this could happen because the immediately
preceding request explicitly handles the case of missing arguments,
but does not bail out, only warns about it and allows the program
to proceed into the code that was changed.  So why were empty or
missing arguments never tested?